Contest/requirements.txt
torch
torchvision
pandas
pillow
numpy
scikit-learn